refactor: move utility folder
